Changed to int to uint to avoid overflow

+++ src/Todoist.Net/Models/ItemState.cs
@@ -13,7 +13,7 @@
         /// <param name="indent">The indent.</param>
         /// <param name="inHistory">The in history.</param>
         /// <param name="order">The order.</param>
-        public ItemState(ComplexId itemId, bool isChecked, int indent, bool inHistory, int order)
+        public ItemState(ComplexId itemId, bool isChecked, uint indent, bool inHistory, uint order)
             : base(itemId, order, indent)
         {
             IsChecked = isChecked;
@@ -32,13 +32,13 @@
         /// <value><c>true</c> if this instance is checked; otherwise, <c>false</c>.</value>
         public bool IsChecked { get; }
 
-        internal override int[] ToArray()
+        internal override uint[] ToArray()
         {
-            var array = new int[4];
+            var array = new uint[4];
 
-            array[0] = InHistory ? 1 : 0;
+            array[0] = (uint) (InHistory ? 1 : 0);
 
-            array[1] = IsChecked ? 1 : 0;
+            array[1] = (uint) (IsChecked ? 1 : 0);
 
             array[2] = Order;
